Bug #53373 Comments are not properly removed from the            pre-processed query buffer.
Preprocessing comments with embedded SQL, the "/*!<sql>*/"
style, failed to provide correct SQL.  The preprocessed
output is used in the statement replication subsystem,
and in some cases, the SQL statement sent to the slave was
not functionally equivalent to the one executed on the
master. 

The C (and C++) programming language states that a
long ("/* ... */") comment in the input should be 
preprocessed and replaced by a single space character 
in the output.   This behaviour is also adopted by this
fix.  To reduce the number of affected test cases, this
space is only inserted when either the preceding or following
character is not a whitespace (i.e a whitespace duplication
removal)