Bug #58109: Function my_load_defaults() does not read             correct file
When passed an absolute filename, my_load_defaults()
is supposed to read only that file, but instead it
reads values from the defaults file found in a previous
execution of the function (which is saved in
my_defaults_file).

This patch fixes this by changing order of checks in
my_search_option_files() so that it first checks if an
absolute path was supplied, and in that case uses that
file only.