BUG#58160: Incorrect behaviour of rpl_chage_master
rpl_change_master was not resetting the master_retry_count
at the end of the test. Thence it would leave retry_count
with a huge value, which could cause other tests executing
after it to time out.

We fix this by resetting the value before the 
rpl_change_master test ends.