Fix for Bug#58243 RQG test optimizer_subquery causes server crash when running with ICP
The crash was due to hitting an assert in InnoDB that checked that the
same transaction could only have one active data access to InnoDB. In
these cases there were two ongoing data accesses by the same
transaction. This happened because the server was pushing down an
index condition that contained a subquery. When InnoDB executed the
index condition call back function this would result in a second call
to InnoDB due to executing the subquery from within the index conditon
function.

To avoid this problem we should not push down index conditions that
contains an subquery. The existing code for determining which part of
a table's where condition that can be pushed down already had code for
handling this but due to inconsistent data in the condition's item
tree (or wrong use of methods on it) this code did not work as
expected. Two cases where this happened (see the code in
uses_index_fields_only()):

1. Test for constant items:

    if (item->const_item())
       return TRUE;

   If the item tree contains a subquery the call to const_item() 
   could in some cases return true even when the subquery has to
   be executed later. In this case we would include the subquery
   in the pushed index condition.

   Fix for this problem: Check the item's with_subselect field.
   If this is true then do not include it, other let the item
   tree be included.

2. Test for which tables the item tree contains:

    if (!(item->used_tables() & tbl->map))
       return other_tbls_ok;

   This code would accept to include an item tree if it did not
   use the table we are using for push down. If the item tree
   contained a subquery then the used_tables() method can return
   the wrong set of tables and this could result in that the item
   was wrongly included in the condition that was pushed down.

   Fix for this problem: Remove this test. This will result in that we
   will potentially recursively traverse more of the item tree.  The
   evaluation will be done by the main switch statement and the
   decision about whether to accept accesses to data from other tables
   will be done on the field item.

Two test cases are included that would trigger the two cases.