WL#5692 "Enable Disk-Sweep Multi-Range Read optimizer feature".
This patch contains the second part of this worklog. With this change
the MySQL server will be built with support for the Multi-Range Read
(MRR) feature. The default will be that cost-based MRR will be
enabled.

The following changes to MySQL are implemented:
      
1. Enable disk-sweep Multi-Range Read for MyISAM and InnoDB
        
With this feature enabled the optimizer will consider to use the
disk-sweep multi-range read strategy when reading a range of
records using a secondary index. The decision whether to use
MRR or not is (by default) cost-based. The disk-sweep MRR strategy
is implemented for MyISAM and InnoDB.
      
The main change by enabling MRR should be improved performance for
range queries that needs to access records from the base table using a
secondary index when the base table requires disk-access. The user
should be able to see if MRR has been used by looking in the "Extra"
column of the EXPLAIN output. If this field contains the string "Using
MRR" then MRR will be used.
      
2. Enable the user to control the use of Multi-Range Read
      
When MRR has been enabled the user will be able to control the use 
of MRR by using two optimizer_switch configuration variables:

* mrr: if this is on then MRR is available for use by the optimizer
  The default value is "on".

* mrr_cost_based: This determines whether the use of MRR should be
  based on cost estimates done by the optimizer (value: on) 
  or if MRR should always be used (value: off). The default value
  is "on".