SPJ-sca-scan: Cherry picked fix for bug#58818: Incorrect result for IN/ANY subquery with HAVING condition
The fix is to ensure that the HAVING condition 'IS TRUE' when doing a
Item_in_subselect::single_value_transformer

If the ::single_value_transformer() find an existing HAVING condition it used
to do the transformation:

   1) HAVING cond -> (HAVING Cond) AND (cond_guard (Item_ref_null_helper(...))

As the AND condition above is Mc'Carty evaluated, the idea was that the
right side of the AND cond would be executed only if the HAVING evaluated to true.

This logic failed to considder the 3-value logic required by NULL/UNKNOWN evaluation.
Evaluating 'UNKNOWN AND <right cond>' requires the right AND-argument to be evaluated and
'false' returned if <right cond> evaluated to false, else unknown.

To get the required behaviour this fix change the transformation in 1) to be:

   2) HAVING cond -> ((HAVING Cond) IS TRUE) AND (cond_guard (Item_ref_null_helper(...))