Remove stats locking from ha_innobase::info()
This is a workaround for
"Bug#58974 ha_innobase:info() tries to lock some latches even if
HA_STATUS_NO_LOCK is given"
to get into 5.6.1 because those locks are hurting performance badly.

A better solution which does not allow inconsistent stats to be read
into ha_innobase::stats is being baked in rb://540. Those locks were
introduced recently. Before that the race has always exited. It is not a
big problem to read inconsistent stats for a single query. The stats are
just an estimates.