Bug #58996 SHOW CREATE TRIGGER fails when subject table is shadowed by            a temporary table.
The problem was that SHOW CREATE TRIGGER failed to show a table trigger
if a temporary table with the same name as the trigger table existed.
This even if temporary tables cannot have triggers and SHOW CREATE
TRIGGER specifies the name of the trigger (not the name of the table).

This patch fixes the problem by forcing SHOW CREATE TRIGGER to ignore
temporary tables when trying to open the table for which the trigger
is defined.

Test case added to show_check.test.