Backport to mysql-5.1-telco-7.0:
Fix for bug#58626, Incorrect result for WHERE <column> IN (<subquery>) IS UNKNOWN.
      
NOTE: This fix is a backport from 5.6.x which already seems to have fixed this problem.
      
The fix ensures that if 'Full scan on NULL key' access method may be used
for a table, we can't assume that any part of the predicate is covered by
the REF-key. (The join_tab is known to have 'Full scan on NULL key' if
any cond_guards[] has been defined for 'join_tab->ref')
      
part_of_refkey() will therefore return '0' if a potential'Full scan on NULL key' 
is detected - Which will force make_cond_for_table() to include all part of a 
predicate covering the specified 'tables' and 'used_table' mask.