Bug#58629 - Disable UNIV_MUST_NOT_INLINE InnoDB disables inlining of code for all platforms that are not Linux and Sun;
# if !defined(__GNUC__) && !(defined(__SUNPRO_C) || defined(__SUNPRO_CC))
#  undef  UNIV_MUST_NOT_INLINE
#  define UNIV_MUST_NOT_INLINE
# endif

This patch and bug fix enables inlining of code on all other platforms.
It removes all references to UNIV_MUST_NOT_INLINE in the build files,
but keeps the #ifdef UNIV_MUST_NOT_INLINE in univ.i and the UNIV_NONINL
in various files. This is because we may want to disable the inlining of .ic files
to troubleshoot some future problem.

RB://538  Approved by Marko Mäkelä