BUG#11747577 Not able to recover binary/blob data correctly using mysqlbinlog
The queries generated by mysqlbinlog could not be executed by 'mysql' client
if there was any 0x00 in it. Or the queries recovered diverged data if there
was any 0x0D0A sequence in it. The reason is that 'mysql' client always
translated '0x0D0A'('\r\n') to '0x0A'('\n') and treated '0x00' as the end
of a query if it is in non-interactive mode(receiving queries from a file
or a pipe).
            
This patch added the option 'binary-mode' in 'mysql' program.  'mysql' will
not translates '0x0D0A' to '0x0A' and not treat '0x00' as the end of a query
if binary-mode is set 1 and it is in non-interactive mode.  Meanwhile all
mysql client commands except 'delimiter' and '\C' are disabled. This will
improve the performance when mysql applies huge logs.