Fix for bug#11804277 - INCORRECT INDEX MAY BE SELECTED DUE TO INSUFFICIENT STATISTICS FROM CLUSTER
Add heuristics to ha_ndbcluster::records_in_range()
which identifies a range as:

 - An open bound range  (<column> LT/GT <bound)
 - A closed bound range (<column> BETWEEN <low> AND <HIGH>)
 - A (partial) EQ-range (<column> EQ <bound>)

... Or a combination of these....

These are handled as follows:

Open bound ranges
-----------------
Without a histogram of how the values in the index are
distributed, we can only assume an equal distrubution.
A statistically correct estimate for a condition of the form
'<column> LT/GT <bound>' would then have been to assume it selects
50% of the rows in the table. However, I have experienced that
this will cause the range-cost to directly compete with the cost
of a full table scan. We should therefore be somewhat more
conservative and estimate 10% of the rows to be returned.

Closed bound range
------------------
We assume this to be somewhat better than an open bounded range
returning 5% of the rows in the table.

EQ-range
--------
An EQ-range will excatly specify a fraction of the first part of
an index. It is reasonable to assume:

  - Specifing a larger fraction of the index will improve the
    selectivity of the EQ-range.
  - Each part of the specified EQ-range will have the same
    selectivity.

We can model this as a Binomial Distribution of the indexed
values. http://en.wikipedia.org/wiki/Binomial_distrib