Initial patch for Bug#12321461 CRASH IN DSMRR_IMPL::DSMRR_INIT ON SELECT STRAIGHT_JOIN
This crash is caused by a too strict assert that was added in the patch
for Bug#58463. This assert crashes the server when using MyISAM and
the query uses ICP on a primary index in combination with MRR.

As an initial fix to avoid that this assert results in a lot
of unnecessary crashes during testing this patch disables the assert.