Bug#11822517: Refactor call interface of choose_plan()
Part 2 - Change meaning of greedy_search() argument 'remaining_tables'.

The purpose of this patch is to make the interface of greedy_search()
and related functions simpler.
It is also a preparatory step for a simpler test for dependencies when
doing semi-join materialization together with outer join (WL#5561).

The meaning of the argument remaining_tables to greedy_search() is changed:
Before this patch, it meant all tables not yet added to the current
join plan, regardless of whether the tables were to be optimized or not.
After the patch, it means all tables remaining to be added, but not
including tables that will not be optimized.

The difference is important when making a plan for a materialized
semi-join nest. Before, remaining_tables included also the tables
belonging to the query, but not part of the semi-join nest being
optimized. This makes for simpler implementation of
best_extension_by_limited_search() and easier maintenance.

However, best_access_path() requires that information about all tables
not yet added to the plan be available. We implement this by adding
a member variable excluded_tables in the Optimize_table_order class,
which contains the set of all tables in the query block that are not
to be optimized in the current operation. Notice that excluded_tables
is passed to best_access_path() only where it can be non-empty.

sql/sql_select.cc
  New member variable excluded_tables is added to Optimize_table_order.

  In choose_table_order(), join_tables is calculated as the set of
  tables to be optimized.

  In best_extension_by_limited_search(), use of allowed_tables is
  eliminated, and logic has been slightly simplified.