Bug 12384651 - ASSERT AT LOCK_TABLE_QUEUE_VALIDATE ON RQG TEST RUNS
Remove the assertion and accompanying code. We were counting the table
locks held by a transaction multiple times. This is because a transaction
can have multiple table locks. The stale locks are not removed and this
is by design. We are already checking whether the transaction holds the
table lock that the table lock list claims that it does. What we were
trying to check for was whether there are any locks in the trx table lock
vector that the table doesn't know about.

This is a debug change, should not affect production code.