WL#4896 "Performance Schema Net IO"
Fix PB2 crash:

For client connections, set socket thread owner only after the socket
has been relegated to a client thread. Otherwise, multiple new client
sockets will share the listening thread as the thread owner, causing
a race condition when writing to the events waits buffer.

For the listening sockets, set the thread owner on the main thread
and again when the socket is relegated to a listening thread.