Bug#11747690 (former 33650) MYSQL_ALTER_TABLE() UNNECESSARILY DOES FULL TABLE COPY
The problem was that full table copy was done when creating an
index on a UTF-8 VARCHAR column (rather than using in-place
ALTER TABLE).

This problem is not repeatable using InnoDB. It is repeatable
using MyISAM, but MyISAM does not support in-place add index in
general, not just for UTF-8 VARCHAR columns. Therefore the current
behavior is not a bug with regards to MyISAM.

This patch adds a regression test for InnoDB and contains no
code changes.