Bug#11762345 54927: DROPPING AND ADDING AN INDEX IN ONE              COMMAND CAN FAIL IN INNODB PLUGIN 1.0
If a single ALTER TABLE statement drops and adds an index
with the same name, the operation has to be done using full
table copy rather than changing the table in-place. This
problem was partly fixed by an index name check added in
the patch for Bug#49838.

However, a problem remained if the new index was not given
an explicit name. In such cases the index is automatically
given the same name as the first column in the index defintion.
If this name happened to be the same name as the index to be
dropped, the match would not be detected by the fix for
Bug#49838. This caused ALTER TABLE to fail with "Incorrect
index name".

This patch fixes this problem by postponing the index name
check until after unnamed new indexes have been given a
name. This allows the check to catch situations described
above and force the use of full table copy (rather than
aborting with "Incorrect index name").

Test case added to innodb.innodb_mysql.test.