WL5353 Primary Cluster Conflict Resolution
Include NdbApi transaction ids with data change events.

A new method, getTransId() is added to the NdbEventOperation class.
For data change events (INSERT, UPDATE, DELETE), this method returns the
transaction id of the NdbApi transaction which caused the event.  This
is the same transaction id as returned by NdbTrasaction::getTransactionId()
when executing the transaction.

Existing NdbApi event testcases are augmented to verify that the transaction
id is passed through the cluster as expected.