Bug#11765699 (58690) !TABLE ||                     (!TABLE->READ_SET ||                      BITMAP_IS_SET(TABLE->READ_SET, FIELD_INDEX
JOIN::conds points to the same expression tree as 
SELECT_LEX::where, so modifying the expression tree of 
JOIN::conds effectively does the same modifications to 
SELECT_LEX::where.

JOIN::conds is optimized in JOIN::optimize()

  conds= optimize_cond(conds);

optimize_cond() makes modifications to the conds input 
parameter. One of these modifications removes an item from 
the input conds parameter. This was not a problem for 
JOIN::conds because removed items are included in the 
condition returned from the function and therefore assigned 
to conds as can be seen above. However, the removed parts of
the condition were lost from SELECT_LEX::where.

This bug was about a query with three levels of subselects:

SELECT 1 FROM t1 WHERE a =
  (SELECT 1 FROM t2 WHERE b =
     (SELECT 1 FROM t1 t11 WHERE c = 1 OR t1.a = 1 AND 1 = 2)
   GROUP BY b
);

Prior to optimization, select#3 has

join->conds = select_lex->where = "c = 1 OR t1.a = 1 AND 1 = 2"

As part of optimize_cond(), remove_eq_conds() sees that
"t1.a = 1 AND 1 = 2" is FALSE and correctly removes this 
condition. This leaves the OR item with only one argument: 
"c = 1". With only one argument left in this OR, 
internal_remove_eq_conds() empties the argument list of the OR 
and returns "c = 1". The effect is that after optimize_cond(), 
select#3 has:

join->conds = "c = 1"
select_lex->where = ""   // <- !!

Fixed by not emptying the argument list of an Item_cond even 
when there is only one argument. The effect is that after 
optimize_cond(), select_lex->where is an OR-item with only 
one argument: "c = 1". This is slightly inefficient (evaluation
would have to go through the OR), but it works correctly. Note 
that condition evaluation is performed using the correctly 
optimized JOIN::conds.