This commit adds a regression test for http://lists.mysql.com/commits/138223, which is:
"This is a fix for an error that would occur when pushing an index scan that is sorted on a string field.
The problem was an assumption that compare_ndbrecord() would always return 0,-1 or 1, when it can in fact return any
number. This could cause results to be sorted in the wrong order, or mysqld to loop for ever."